Fixing random_spit() TypeError

+++ convnet.py
@@ -139,7 +139,7 @@ SPECTR_DIR = "dataset/content/fma"
 
 dataset = customDataset(annotation_file=CSV_DIR, data_dir=SPECTR_DIR, transform=composed_transform)
 dataset_length = len(dataset)
-train_dataset, validation_dataset, test_dataset = torch.utils.data.random_split(dataset, [0.8*dataset_length, 0.1*dataset_length, 0.1*dataset_length])  # using 80/10/10% (training/validation/test) ratio
+train_dataset, validation_dataset, test_dataset = torch.utils.data.random_split(dataset, [int(0.8*dataset_length), int(0.1*dataset_length), int(0.1*dataset_length)])  # using 80/10/10% (training/validation/test) ratio
 
 train_data_loader = DataLoader(dataset=train_dataset, batch_size=batch_size, shuffle=True)
 validation_data_loader = DataLoader(dataset=validation_dataset, batch_size=batch_size, shuffle=True)
